Rename to cli package
